requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.8.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.0.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
databricks-sdk==0.46.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.110.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipinfo==5.1.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.21.0
mlflow-skinny==2.21.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
nvidia-nccl-cu12==2.26.2
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.7.1
pydantic_core==2.20.1
Pygments==2.19.1
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
secure==0.3.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.36.3
threadpoolctl==3.5.0
tldextract==5.1.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==3.0.0
yarl==1.18.3
zenml==0.75.0
zipp==3.21.0
